Lots of fixes & improvements
