+++ Cargo.toml
@@ -40,7 +40,7 @@ futures = { version = "0.3.8", optional = true }
 getrandom = "0.2.0"
 log = "0.4.11"
 maybe-async = "0.2.1"
-serde = { version = "1.0.115", features = ["derive"] }
+serde = { version = "1.0.115", default-features = false }
 serde_json = "1.0.57"
 thiserror = "1.0.20"
 url = "2.2.2"
