+++ rspotify-model/src/context.rs
@@ -3,7 +3,8 @@
 use super::device::Device;
 use super::PlayableItem;
 use crate::{
-    millisecond_timestamp, option_duration_ms, CurrentlyPlayingType, DisallowKey, RepeatState, Type,
+    custom_serde::{millisecond_timestamp, option_duration_ms},
+    CurrentlyPlayingType, DisallowKey, RepeatState, Type,
 };
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Deserializer, Serialize};
